requirements.txt
# Core dependencies
Pillow>=9.0.0
imagehash>=4.0.0
flask>=2.0.0
numpy>=1.20.0           # Required by imagehash
pillow-heif>=0.10.0     # HEIC/HEIF format support
scikit-learn>=1.0.0     # K-means clustering for color sorting
piexif>=1.0.0           # EXIF metadata manipulation
tqdm>=4.0.0             # Progress bars